+++ clean_python/s3/__init__.py
@@ -1,3 +1,6 @@
 from .key_mapper import *  # NOQA
+from .s3_bucket_options import *  # NOQA
 from .s3_gateway import *  # NOQA
 from .s3_provider import *  # NOQA
+from .sync_s3_gateway import *  # NOQA
+from .sync_s3_provider import *  # NOQA