+++ clean_python/s3/sync_s3_provider.py
@@ -3,32 +3,30 @@
 import logging
 from typing import TYPE_CHECKING
 
-try:
-    import aioboto3
-except ImportError:
-    aioboto3 = None
+import boto3
 from botocore.client import Config
-
-from clean_python import Provider
 
 from .s3_bucket_options import S3BucketOptions
 
 if TYPE_CHECKING:
-    from types_aiobotocore_s3.client import S3Client
+    from mypy_boto3_s3.client import S3Client
 
-__all__ = ["S3BucketOptions", "S3BucketProvider"]
+__all__ = ["S3BucketOptions", "SyncS3BucketProvider"]
 
 logger = logging.getLogger(__name__)
 
 
-class S3BucketProvider(Provider):
+class SyncS3BucketProvider:
     def __init__(self, options: S3BucketOptions):
         self.options = options
 
-    async def connect(self) -> None:
-        assert aioboto3 is not None
-        self._session = aioboto3.Session()
-        self._client = await self._session.client(
+    @property
+    def bucket(self) -> str:
+        return self.options.bucket
+
+    @property
+    def client(self) -> "S3Client":
+        return boto3.client(
             "s3",
             endpoint_url=self.options.url,
             aws_access_key_id=self.options.access_key,
@@ -43,15 +41,4 @@ class S3BucketProvider(Provider):
                 },
             ),
             use_ssl=self.options.url.startswith("https"),
-        ).__aenter__()
-
-    async def disconnect(self) -> None:
-        await self._client.close()
-
-    @property
-    def bucket(self) -> str:
-        return self.options.bucket
-
-    @property
-    def client(self) -> "S3Client":
-        return self._client
+        )